// task-manager-addon/src/types/interfaces/listings/TasksOrders.ts
/**
 * TasksOrders
 */
export enum TasksOrders {
  Unspecified = 'Unspecified',
  LastActivityAsc = 'LastActivityAsc',
  LastActivityDesc = 'LastActivityDesc',
  MyLastActivityAsc = 'MyLastActivityAsc',
  MyLastActivityDesc = 'MyLastActivityDesc',
  PriorityAsc = 'PriorityAsc',
  PriorityDesc = 'PriorityDesc',
}